Generalized goToIndexPage method from MainActivity into MuistiActivity
